Remove non-standard BOMs from C++ source files
BOMs (byte order markers) were originally designed for UTF-16 and UTF-32
to be able to deal with both big endian and little endian variants. To
this end, the value 0xFEFF is written as a 16bit or 32bit word at the
start of a text file. By inspecting the first bytes of the file, a text
editor can then determine if the file was written in little or big
endian mode.

For UTF-8, this was never necessary or useful. However, there is still a
BOM standardized for UTF-8, to be able to explicitly mark a file as
being UTF-8 encoded; for this the *byte* sequence EF BB BF has to be at
the start of the file. However, using a UTF-8 BOM is generally not
recommended. Alas, Microsoft tools for some reason insist on it
(presumably due to the historic use of UTF-16 (resp. UCS-2) as default
encoding in Windows.

What is never correct, however, is to put UTF-16 BOMs into file using
UTF-8 encoding. Thus this patch removes non-standard UTF-16 BOMs 0xFEFF
from several source files.